Moved traversal matchers out of cypher-compiler-2.3 to remove kernel dependency
