+++ community/cypher/cypher-compiler-2.3/src/main/scala/org/neo4j/cypher/internal/compiler/v2_3/executionplan/builders/TraversalMatcherBuilder.scala
@@ -23,7 +23,7 @@ import org.neo4j.cypher.internal.compiler.v2_3.commands._
 import org.neo4j.cypher.internal.compiler.v2_3.commands.expressions.{IdFunction, Identifier}
 import org.neo4j.cypher.internal.compiler.v2_3.commands.predicates.{Equals, Predicate}
 import org.neo4j.cypher.internal.compiler.v2_3.executionplan.{ExecutionPlanInProgress, PlanBuilder}
-import org.neo4j.cypher.internal.compiler.v2_3.pipes.matching.{BidirectionalTraversalMatcher, MonoDirectionalTraversalMatcher, Trail, TraversalMatcher}
+import org.neo4j.cypher.internal.compiler.v2_3.pipes.matching.{Trail, TraversalMatcher}
 import org.neo4j.cypher.internal.compiler.v2_3.pipes.{EntityProducer, PipeMonitor, SingleRowPipe, TraversalMatchPipe}
 import org.neo4j.cypher.internal.compiler.v2_3.spi.PlanContext
 import org.neo4j.cypher.internal.compiler.v2_3.symbols.SymbolTable
@@ -96,22 +96,22 @@ class TraversalMatcherBuilder extends PlanBuilder with PatternGraphBuilder {
     old ++ solvedPreds.map(_.solve)
   }
 
-  private def chooseCorrectMatcher(end:Option[String],
-                           longestPath:LongestTrail,
-                           startNodeFn:EntityProducer[Node],
-                           startToken:QueryToken[StartItem],
-                           unsolvedItems: Seq[QueryToken[StartItem]],
-                           ctx:PlanContext): (TraversalMatcher,Seq[QueryToken[StartItem]]) = {
+  private def chooseCorrectMatcher(end: Option[String],
+                                   longestPath: LongestTrail,
+                                   startNodeFn: EntityProducer[Node],
+                                   startToken: QueryToken[StartItem],
+                                   unsolvedItems: Seq[QueryToken[StartItem]],
+                                   ctx: PlanContext): (TraversalMatcher, Seq[QueryToken[StartItem]]) = {
     val (matcher, tokens) = if (end.isEmpty) {
-      val matcher = new MonoDirectionalTraversalMatcher(longestPath.step, startNodeFn)
+      val matcher = ctx.monoDirectionalTraversalMatcher(longestPath.step, startNodeFn)
       (matcher, Seq(startToken))
     } else {
       val (endToken, endNodeFn) = identifier2nodeFn(ctx, end.get, unsolvedItems)
       val step = longestPath.step
-      val matcher = new BidirectionalTraversalMatcher(step, startNodeFn, endNodeFn)
+      val matcher = ctx.bidirectionalTraversalMatcher(step, startNodeFn, endNodeFn)
       (matcher, Seq(startToken, endToken))
     }
-    (matcher,tokens)
+    (matcher, tokens)
   }
 
   def identifier2nodeFn(ctx:PlanContext, identifier: String, unsolvedItems: Seq[QueryToken[StartItem]]):
